Don't let non-default-export legacy namespaces clobber legacy sub-namespaces.
For example, in a scenario where two legacy namespaces a.b.c and a.b are loaded in that order, and a.b is not a default export (meaning it never overwrites `exports`), a.b.c is now copied onto a.b when the latter is loaded. In the past, loading a.b would just clobber a.b.c.

Note that behavior doesn't change if a.b is a default export (the most common case of this is assigning the `exports` of a.b to be a class).

This clobbering happens because of the difference in how goog.provide and goog.module.declareLegacyNamespace assign to a namespace. goog.provide files directly modify the namespace. On the other hand, DLN files export an object (as they are modules), and this object is assigned _to_ the namespace. So the ability to modify a namespace after it was already created implicitly from the loading of a child namespace, without destroying the child namespace, is not present in DLN.

Because declareLegacyNamespaces is explicitly meant to support files that transition from goog.provide to goog.module, it's important that we support this case.

RELNOTES[INC]: declareLegacyNamespace logic has been updated to work more like goog.provide.

PiperOrigin-RevId: 317659866